requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==0.4.6
altair==4.2.2
antlr4-python3-runtime==4.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
blinker==1.6.2
braceexpand==0.1.7
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
-e git+https://github.com/facebookresearch/detectron2.git@c39b2d04891495aeb054c7f932ec906211979586#egg=detectron2
easydict==1.9
entrypoints==0.4
filelock==3.9.0
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
hydra-core==1.1.0
idna==3.4
imageio==2.28.1
imgaug==0.4.0
importlib-metadata==6.6.0
iopath==0.1.9
jax==0.4.10
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
kornia==0.5.0
lazy_loader==0.2
libclang==16.0.0
lightning-utilities==0.8.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.1.0
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.1.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.5.1
portalocker==2.7.0
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-image==0.20.0
scikit-learn==0.24.2
scipy==1.10.1
-e git+https://github.com/geekyutao/inpaint-anything@125aa0c342945b75d1adb6946324207848a79aca#egg=segment_anything&subdirectory=segment_anything
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.22.0
streamlit-drawable-canvas==0.9.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchmetrics==0.11.4
torchvision==0.15.2+cpu
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.13
validators==0.20.0
watchdog==3.0.0
webdataset==0.2.48
Werkzeug==2.3.4
wldhx.yadisk-direct==0.0.6
wrapt==1.14.1
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
